chore: move test files
